+++ lib/mpoption.m
@@ -59,6 +59,9 @@ function opt = mpoption(varargin)
 %       [ 'FDXB' - Fast-Decoupled (XB version)                              ]
 %       [ 'FDBX' - Fast-Decoupled (BX version)                              ]
 %       [ 'GS'   - Gauss-Seidel                                             ]
+%       [ 'PQSUM'- Power Summation method (radial networks only)            ]
+%       [ 'ISUM' - Current Summation method (radial networks only)          ]
+%       [ 'YSUM' - Admittance Summation method (radial networks only)       ]
 %   pf.tol                  1e-8        termination tolerance on per unit
 %                                       P & Q mismatch
 %   pf.nr.max_it            10          maximum number of iterations for
@@ -67,6 +70,12 @@ function opt = mpoption(varargin)
 %                                       fast decoupled method
 %   pf.gs.max_it            1000        maximum number of iterations for
 %                                       Gauss-Seidel method
+%   pf.radial.max_it        20          maximum number of iterations for
+%                                       radial power flow methods
+%   pf.radial.vcorr         0           perform voltage correction procedure
+%                                       in distribution power flow
+%       [  0 - do NOT perform voltage correction                            ]
+%       [  1 - perform voltage correction                                   ]
 %   pf.enforce_q_lims       0           enforce gen reactive power limits at
 %                                       expense of |V|
 %       [  0 - do NOT enforce limits                                        ]
@@ -521,6 +530,10 @@ if have_opt0
             end
             if opt0.v <= 11         %% convert version 11 to 12
                 opt0.opf.use_vg = opt_d.opf.use_vg;
+            end
+            if opt0.v <= 12         %% convert version 12 to 13
+                opt0.radial.max_it  = opt_d.radial.max_it;
+                opt0.radial.vcorr   = opt_d.radial.vcorr;
             end
             opt0.v = v;
         end
@@ -1446,6 +1459,9 @@ if ~isstruct(opt)
                 'max_it',               30  ), ...
             'gs',                   struct(...
                 'max_it',               1000  ), ...
+            'radial',               struct(...
+                'max_it',               20   , ...
+                'vcorr',                 0  ), ...
             'enforce_q_lims',       0   ), ...
         'cpf',                  struct(...
             'parameterization',     3, ...
@@ -1537,7 +1553,7 @@ optt = opt;
 %% globals
 %%-------------------------------------------------------------------
 function v = mpoption_version
-v = 12;     %% version number of MATPOWER options struct
+v = 13;     %% version number of MATPOWER options struct
             %% (must be incremented every time structure is updated)
             %% v1   - first version based on struct (MATPOWER 5.0b1)
             %% v2   - added 'linprog' and 'quadprog' fields
@@ -1560,6 +1576,7 @@ v = 12;     %% version number of MATPOWER options struct
             %%        'nose_tol', 'p_lims_tol' and 'q_lims_tol',
             %%        removed option 'cpf.user_callback_args'
             %% v12  - added option 'opf.use_vg'
+            %% v13  - added 'pf.radial.max_it', 'pf.radial.vcorr'
 
 %%-------------------------------------------------------------------
 function db_level = DEBUG