+++ lib/t/test_matpower.m
@@ -47,6 +47,7 @@ tests{end+1} = 't_mips';
 tests{end+1} = 't_qps_matpower';
 tests{end+1} = 't_miqps_matpower';
 tests{end+1} = 't_pf';
+tests{end+1} = 't_pf_radial';
 tests{end+1} = 't_cpf';
 tests{end+1} = 't_islands';
 tests{end+1} = 't_opf_model';