+++ CHANGES.md
@@ -12,6 +12,22 @@ For change history for [MOST][3], see [most/CHANGES.md](most/CHANGES.md).
 Since last release
 ------------------
 
+#### 2/9/17
+  - Add three new power flow algorithms for radial distribution
+    systems selected via the three new options for `pf.alg`, namely
+    `'PQSUM'`, `'ISUM'`, `'YSUM'`. Also includes new MATPOWER options
+    `pf.radial.max_it` and `pf.radial.vcorr`. See Section 4.3 on
+    "Distribution Power Flow" in the manual for details.
+    *Thanks to Mirko Todorovski.*
+  - Add 6 new radial distribution system cases.
+    - `case4_dist`
+    - `case18`
+    - `case22`
+    - `case69`
+    - `case85`
+    - `case141`
+  *Thanks to Mirko Todorovski.*
+
 #### 1/26/17
   - Add sources for MATPOWER User's Manual to `docs/src`.
   - Move MOST User's Manual from `docs` to `most/docs`.